chore: Move releases tests to test directories
