+++ static/js/publisher/pages/Releases/actions/__tests__/currentTrack.test.js
@@ -1,4 +1,4 @@
-import { SET_CURRENT_TRACK, setCurrentTrack } from "./currentTrack";
+import { SET_CURRENT_TRACK, setCurrentTrack } from "../currentTrack";
 
 describe("setCurrentTrack actions", () => {
   const track = "test";
